Bug fixing

+++ org.kevoree.microsandbox.monitoring/src/main/java/org/kevoree/monitoring/ranking/ComponentRankerFunctionFactory.kt
@@ -30,7 +30,7 @@ public object ComponentRankerFunctionFactory {
                 (a,b) ->
                 val x = ComponentsInfoStorage.instance.getExecutionInfo(a.path() as String)
                 val y = ComponentsInfoStorage.instance.getExecutionInfo(b.path() as String)
-                if ((x?.getFailures() as Long) < y?.getFailures() as Long)
+                if ((x?.getFailures() as Int) < y?.getFailures() as Int)
                     -1
                 else
                     1
